added and completed a test file for all the generators in the generators.py file
