--- src/generators_2.py
@@ -1,79 +1,0 @@
-# Напишите генератор, который принимает список словарей и
-# возвращает описание каждой операции по очереди. То есть, тут надо будет бы
-# написать генераторное выражение – это уже от меня.
-# Пример вызова функции:
-# descriptions = transaction_descriptions(transactions):
-# for _ in range(5):
-#     print(next(descriptions))
-# Пример вывода:
-# Перевод организации
-# Перевод со счета на счет
-# Перевод со счета на счет
-# Перевод с карты на карту
-# Перевод организации
-
-# список словарей на вход этой функции:
-
-transactions = (
-    [
-        {
-            "id": 939719570,
-            "state": "EXECUTED",
-            "date": "2018-06-30T02:08:58.425572",
-            "operationAmount": {"amount": "9824.07", "currency": {"name": "USD", "code": "USD"}},
-            "description": "Перевод организации",
-            "from": "Счет 75106830613657916952",
-            "to": "Счет 11776614605963066702"
-        },
-        {
-            "id": 142264268,
-            "state": "EXECUTED",
-            "date": "2019-04-04T23:20:05.206878",
-            "operationAmount": {"amount": "79114.93", "currency": {"name": "USD", "code": "USD"}},
-            "description": "Перевод со счета на счет",
-            "from": "Счет 19708645243227258542",
-            "to": "Счет 75651667383060284188"
-        },
-        {
-            "id": 873106923,
-            "state": "EXECUTED",
-            "date": "2019-03-23T01:09:46.296404",
-            "operationAmount": {"amount": "43318.34", "currency": {"name": "руб.", "code": "RUB"}},
-            "description": "Перевод со счета на счет",
-            "from": "Счет 44812258784861134719",
-            "to": "Счет 74489636417521191160"
-        },
-        {
-            "id": 895315941,
-            "state": "EXECUTED",
-            "date": "2018-08-19T04:27:37.904916",
-            "operationAmount": {"amount": "56883.54", "currency": {"name": "USD", "code": "USD"}},
-            "description": "Перевод с карты на карту",
-            "from": "Visa Classic 6831982476737658",
-            "to": "Visa Platinum 8990922113665229"
-        },
-        {
-            "id": 594226727,
-            "state": "CANCELED",
-            "date": "2018-09-12T21:27:25.241689",
-            "operationAmount": {"amount": "67314.70", "currency": {"name": "руб.", "code": "RUB"}},
-            "description": "Перевод организации",
-            "from": "Visa Platinum 1246377376343588",
-            "to": "Счет 14211924144426031657"
-        }
-    ]
-)
-
-
-#пишем саму функцию:
-def trans_descr(list_curr):
-    for i in list_curr:
-        yield i['description']
-
-gen_descr_trans = trans_descr(transactions)
-
-print(next(gen_descr_trans))
-print(next(gen_descr_trans))
-print(next(gen_descr_trans))
-print(next(gen_descr_trans))
-print(next(gen_descr_trans))
